requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comtypes==1.1.14
idna==3.4
PyAudio==0.2.13
pycparser==2.21
pypiwin32==223
pyttsx==1.1
pyttsx3==2.90
pywin32==305
requests==2.28.2
SpeechRecognition==3.10.0
srt==3.5.2
tqdm==4.65.0
urllib3==1.26.15
vosk==0.3.45
websockets==10.4

PyYAML~=6.0